python/dynamic-commands/requirements.txt
absl-py>=0.12.0
asn1crypto>=1.4.0
beautifulsoup4>=4.9.3
boto>=2.49.0
cachetools>=4.2.1
certifi>=2020.12.5
cffi>=1.14.5
chardet>=4.0.0
cryptography>=3.4.6
dataclasses-json>=0.5.2
decorator>=4.4.2
functions-framework==3.*
gcs-oauth2-boto-plugin>=2.7
gcsfs>=0.7.2
google-api-core>=1.26.1
google-api-python-client>=2.0.2
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.3
google-auth>=1.28.0
google-cloud-bigquery>=2.13.0
google-cloud-core>=1.6.0
google-cloud-firestore>=2.0.2
google-cloud-gsuiteaddons>=0.3.1
google-cloud-pubsub>=2.3.0
google-cloud-scheduler>=2.1.1
google-cloud-secret-manager>=2.8.0
google-cloud-service-usage>=1.0.0
google-cloud-storage>=1.36.2
google-crc32c>=1.1.2
google-reauth>=0.1.1
google-resumable-media>=1.2.0
googleapis-common-protos>=1.53.0
grpcio>=1.36.1
httplib2>=0.19.0
idna>=3.3
immutabledict>=2.2.1
inflection>=0.5.1
ipaddress>=1.0.23
lxml>=4.6.3
messytables>=0.15.2
oauth2client>=4.1.3
oauthlib>=3.1.0
opentelemetry-api>=1.0.0rc1
opentelemetry-instrumentation>=0.18b0
protobuf>=3.15.6
pyasn1-modules>=0.2.8
pyasn1>=0.4.8
pycparser>=2.20
pyOpenSSL>=20.0.1
python-card-framework>=1.0.0
python-jose>=3.2.0
python-oauth-token-manager>=0.1.*
python-service-builder>=0.1.0
pytz>=2021.1
pyu2f>=0.1.5
ratelimit>=2.2.1
requests-oauthlib>=1.3.0
requests>=2.27.1
retry-decorator>=1.1.1
rsa>=4.7.2
six>=1.15.0
SocksiPy-branch>=1.1
stringcase>=1.2.0
uritemplate>=3.0.1
urllib3>=1.26.8
web.py>=0.62
yarl>=1.4.2
